- Fixed some bugs with Groza (signed char!) russian font requires unsigned. Changes were made to text.c, etc
- DJGPP now does 'clean' properly.

- other minor fixes.

+++ src/core/graphics.c
@@ -215,7 +215,7 @@ void shake_screen (int n)
 
 
 
-void put_text_character (int l, int x, int y, int c, int fg, int bg)
+void put_text_character (int l, int x, int y, unsigned int c, int fg, int bg)
 {
 	int x1, y1, xx, yy, cc;
 	UINT8 *p;
@@ -225,7 +225,7 @@ void put_text_character (int l, int x, int y, int c, int fg, int bg)
 		c = 1;
 #endif
 
-	p = cur_font + (c * CHAR_LINES);
+	p = cur_font + ((unsigned int)c * CHAR_LINES);
 	for (y1 = 0; y1 < CHAR_LINES; y1++) {
 		for (x1 = 0; x1 < CHAR_COLS; x1 ++) {
 			xx = x + x1;
